+++ packages/mui/src/types.ts
@@ -1,0 +1,21 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+
+// This file is used to define custom types for MUI components to work with the StrataKit customizations.
+// See: https://mui.com/material-ui/customization/theming/#typescript
+// See: https://www.typescriptlang.org/docs/handbook/declaration-merging.html#module-augmentation
+
+import "@mui/material/Button";
+
+declare module "@mui/material/Button" {
+	interface ButtonOwnProps {
+		/**
+		 * The default variant with `@stratakit/mui` is `"contained"`.
+		 *
+		 * @default 'contained'
+		 */
+		variant?: "contained" | "outlined" | "text";
+	}
+}